requirements.txt
surya-ocr==0.4.15
pypdfium2==4.30.0
streamlit==1.37.0
openai==1.35.13
